pyproject.toml
[project]
name = "mcp-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=0.3.8",
    "langchain-mcp-adapters>=0.1.11",
    "langgraph>=0.6.8",
    "mcp-use>=1.3.11",
    "mcp[cli]>=1.16.0",
    "python-dotenv>=1.1.1",
]
